add C++ solution

+++ algorithm/C++/001-two_sums.cpp
@@ -3,7 +3,7 @@ public:
     vector<int> twoSum(vector<int>& nums, int target) {
         unordered_map<int, int> map;
         uint size = nums.size();
-        for(short i=0; i < size; ++i){
+        for(int i=0; i < size; ++i){
             if(map.find(target - nums[i]) != map.end()){
                 return vector<int>{i, map[target - nums[i]]};
             }
